First iteration of minimized developer container image
Got down to 389.2MB, from 600+MB on debian/ubuntu minimalization efforts!
However two issues remain:

1. pre-commit uses shutil.copytree, which dies spectacularly on alpine.
Suspect it is due to symlinks, but the `#RUN make dev` line in the Dockerfile
should replace the `RUN pip3 install --upgrade -e .[dev]` line when this is fixed.

2. There's a test failing, due to coconut throwing segfault when calling the binary.
Should also be looked into. Suspect it could be either (a) a 'real bug', or more likely
(b), a limitation of the minimized alpine build which somehow throws the Coconut binary off.